+++ CHANGES.md
@@ -1,8 +1,11 @@
 # v0.12.0rc
 
+- **new** [scxmlrun-all](tools/scxmlrun-all):
+  run several SCXML processes (along with generic MQTT processes) in parallel.
+
 # v0.11.0 (2018-12-27)
 
-- [new] introduce _"event repeater"_ (invoked with "-r") which transparently transmits incoming events without any interpretation.
+- **new** introduce _"event repeater"_ (activated by: `scxmlrun -r`) which relays incoming events to elsewhere.
 
 - add/update examples
 - add `*.conf` files for testing with [shelltest](https://github.com/simonmichael/shelltestrunner)
@@ -10,7 +13,7 @@
 
 # v0.10.0 (2018-11-08)
 
-- [new] introduce JS functions: `SCXML.raise` and `SCXML.send`
+- **new** introduce JS functions: `SCXML.raise` and `SCXML.send`
 
 - add examples: microwave, calc (with web ui), traffic-control, and blackjack
 - add examples: voting and vulnerable